"Barrel full of Barrels" is now an auto-adventurable location, with a menu in the Choice Advs tab to select which rows of booze you want to collect. (Goals other than booze aren't supported - does anybody actually do that?)
The same probability data as is used by the Barrels decorator is used to
select which barrel to smash, so this feature should be maximally efficient. 
However, the possibilities of an exploding barrel, finding a clover, or
losing to a mimic (none of which help to identify the Barrels pattern) mean
that no guarantee can be made as to how many turns it would take to find all
the booze.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6369 b29ace70-8910-0410-8dcc-aa2fc6433167