Adds online help to the CLI - "help" shows the full command list (long!) with brief descriptions, "help <subject>" shows commands with <subject> in their name or description, and "help help" shows the conventions used in the command descriptions.
There will undoubtedly be commands in the list you weren't aware of, but
these are actually new: "budget", "pull outfit <name>", "condref", and
"moleref".

This feature required a complete rewrite of how commands are handled, so that
usage text could be associated with them; please report any anomalies,
especially in CLI scripts.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6457 b29ace70-8910-0410-8dcc-aa2fc6433167