Allows powders & nuggets to be selected in the Pulverize panel, for either Malus-ing yourself or sending to wadbot.  They appear in the list whenever the corresponding element is selected, and the yield filter includes either "1N" or "1W" as appropriate.
Fixed some problems with using the Pulverize button or context menu item in
other lists with powders or nuggets - that was implemented, but didn't work
with nuggets due to a typo, and would upgrade ALL powders of the selected
type (regardless of number requested) all the way to wads (even if the player
wanted the nuggets).  It should now behave as expected.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6623 b29ace70-8910-0410-8dcc-aa2fc6433167