Initial support for summoning Pastamancer combat entities (ghosts)
We recognize when you summon one, and remember the name and type
We track whenever you use one in battle (which you can do, 10 times per day)
You should be able to say "summon" in a CCS to summon your current ghost
The CLI "entity" command tells you what we know about your current entity.

Note that we don't - quite - track experience correctly: we increment experience
when you summon the combat entity, not when you win the battle with it still
present.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6696 b29ace70-8910-0410-8dcc-aa2fc6433167