Phase 1 of a new feature for tweaking combat behavior without having to constantly edit a CCS.  A Special field has been added to the Adventure panel, between Action and Goals: it contains icons for the enabled special actions, each with a tooltip explaining exactly what it does.  Clicking in this area pops up a menu for selecting and configuring special actions.  The four actions implemented so far are:
* Pickpocket before simple actions - on by default, giving the previous
behavior of automatically trying to steal when it seems appropriate.  Can now
be turned off, for example in Hobopolis where the determination of
appropriateness doesn't work quite the same.

* Cast Entangling Noodles before simple actions - allows a much wider range
of useful strategies to be set without having to create a CCS, and still
allows the benefit of an automatically set auto-attack.

* Minimum poison level for anti-anti-antidote use - this is a sub-menu for
selecting the threshold at which an antidote will be used in combat.  Note
that there is not yet any attempt to ensure that you have an antidote in
inventory; doing that intelligently will require knowing which adventure
locations have poisonous monsters.

* Use MP restorers in combat - if turned off, any skill which you don't have
enough MP to use will simply be skipped.  This will result in combat being
aborted if no following actions are capable of finishing off the monster.

Some of the forthcoming special actions will require specification of the
combat round in which they are to be performed; this will be done
automatically for simple battle actions, or via a "special action" line in a
CCS.  You can add that line to a CCS now, and it will automatically be
present in a newly-created CCS, but it does nothing yet.  Also, there is a
new "skip" action that can be present in a CCS, explicitly doing nothing;
there's no reason to actually enter that in a CCS, but its availability makes
some of the internal workings cleaner.

Clicking "script" in a relay browser battle no longer forces use of a CCS - a
simple battle action constitutes a usable script (and is, in fact, the
strategy the user most recently specified).

Birdform's "look for shiny objects" action is now treated as entirely
equivalent to pickpocketing; it will be used automatically under the same
conditions as pickpocket if a simple battle action is selected, and can be
requested via a "steal" action in a CCS.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6902 b29ace70-8910-0410-8dcc-aa2fc6433167