Monster data can now include the name of a poison effect that can be inflicted on the player.  This goes into the 2nd field of monsters.txt, enclosed in double quotes, separated from other elements in that field by a single space.  Poison data has been added for all monsters listed on the relevant effect pages of the Wiki, and the version number of the file has been bumped.
Adventure areas track the most severe poison that can be inflicted by any
monster found there; if this would trigger an in-combat anti-anti-antidote
use, and the player doesn't have a bezoar ring equipped, adventuring there
will ensure that you have an antidote in inventory.

Added antidotes to the list of items that are purchased in bulk when not
under ronin/hardcore restrictions.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@6929 b29ace70-8910-0410-8dcc-aa2fc6433167