Provide useless option to not strip linebreaks when logging HTML in debug log.
Using items in-line was erroneously trimming what was displayed to what was
between the first <table> tag and the first </table> tag. That breaks when the
output contains nested tables and/or multiple tables.

It appears to be completely unnecessary, in addition to being simply wrong, so
it's gone.

You may to remove data/relay/basics.js to get ths fix.e


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@7059 b29ace70-8910-0410-8dcc-aa2fc6433167