Bottles of sake can now be made by fermenting white rice.
Added a mechanism for overriding the expected number of turns taken by an
adventure request, for cases where the caller has some special knowledge
about what's going to happen.  Used this to specify that the visit to Noob
Cave (to get the Welcome Back adventure after finishing a gong) will take 0
turns, so that it doesn't inappropriately trigger counters.

Added new ASH function for using this mechanism:
boolean adv1(location loc, int turnsUsed, string filterFunc)
This will adventure exactly once in the specified location (even if it took
no turns), although it will follow chained choice adventures for as long as
you have preset options.

turnsUsed overrides the normal number of adventures that are expected to be
consumed at this location, or -1 can be passed to use the default. 
Specifically, 0 will prevent any counters from triggering - of course, if an
adventure actually is consumed, any counters that you kept from triggering
will be lost.

filterFunc is the name of a combat filter function defined elsewhere in your
script, with the same behavior as in the 3-paramater version of adventure(). 
An empty string can be passed to use your battle action or CCS as normal.

Your autoattack is now reset when adventuring with a combat filter active,
just as if you were using a CCS.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@7434 b29ace70-8910-0410-8dcc-aa2fc6433167