Revamp how adventures get logged in the gCLI and session log. Used to be, they got logged just before a URL was submitted: we'd figure out where you were trying to go, do some area specific preparation (like resetting autoattack), and go. Later, when the response came back, we'd look at the response and figure out what the Encounter was.
The problem with that is that KoL has many ways of either denying an adventure
outright or sending you to a different area entirely. So, in the new scheme, we
examine the URL, as before, and decide if it looks like it will be an adventure.
We do some area specific preparation and save what we deduced and go. Later,
when the response comes back, we look at the saved data, the response text, and
the (possibly redirected) URL to decide what really happened. If we really
adventured, we log the location in the gCLI and session log, as before, and go
on to figure out the encounter.

This fixes a lot of things:

- If you were outright denied an adventure because you are out of turns, can't
get to an area, are too drunk to go in the Daily Dungeon or Pyramid Lower
Chambers, etc., nothing is logged.
- If you were redirected from an adventure area because you changed your outfit
and (like mining, or the pirate cove, or the Dwarf Factory COmplex), nothing is
logged.
- If you are redirected to an entirely different area and finish the encounter
without input (trying to adventure in Hippy Camp, but have the outfit, so are
redirected to Hippy Camp in Disguise, for example. Or being redirected to the
Drunken Stupor.), we figure out where you ended up by looking at the Adventure
Again link on the result page.
- If you try to adventure with 0 HP, we no longer get fooled by the response
and list "your campsite" as the encounter because it happened to be bold.

Unfortunately, if KoL redirects you into a choice adventure or a fight (and you
didn't autoattack and win immediately), there is nothing we can do to figure
out where you really ended up. Therefore, this will not notably improve logging
of Teleportitis adventuring. Perhaps I can get some help from KoL itself...


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@7543 b29ace70-8910-0410-8dcc-aa2fc6433167