Fix IDs of evil noodles, nauseating reagent - they were swapped.
There are actually six class-specific versions of the "A Little Bit Evil"
effect - what a mess!  For now, disambiguate them by appending the player's
class to the name, matching how they are listed on the Wiki.  The descID for
the Seal Clubber version hasn't been posted yet, so that one won't work yet.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@7558 b29ace70-8910-0410-8dcc-aa2fc6433167