Added ASH function: float[monster] appearance_rates(location) The returned map will contain the expected apperance rate for the known monsters in the specified location, normally in the range 0..100.  The rate will be zero for bosses and other one-time-only monsters, negative for ultra-rare monsters.
The map will also contain an entry for $monster[none], giving the expected
rate of noncombat encounters.  This will be negative for locations where this
isn't known, in which case the monster encounter rates will be calculated as
if though the location was 100% combats.

The results take combat rate modifiers into account.  Queue effects,
Olfaction, and monster banishers are not considered now, but may be in the
future.  Special mechanics (such as monsters that are unlocked by some
action) aren't considered, and may never be due to difficulty in expressing
their conditions in the data files.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@7572 b29ace70-8910-0410-8dcc-aa2fc6433167