Allow ASH functions invoked via the namespace interpreter to be prompted for parameters (similar to the main method in ASH scripts).  Anything typed into the CLI after the name of the function will be treated as a single string parameter.  Ambiguity is resolved as follows:
If a parameter is provided, choose the function that has a single string parameter.  If no such function is available, but there is exactly one function that needs one parameter, then choose that function.  If no such functions are available, consider the invocation to be ambiguous and abort.

If no parameter is provided, choose the function that has no parameters.  If no such function is available, then choose the function with the fewest number of parameters.  If multiple functions tie for the fewest number of parameters, consider the invocation to be ambiguous and abort.

git-svn-id: https://svn.code.sf.net/p/kolmafia/code@7680 b29ace70-8910-0410-8dcc-aa2fc6433167